+++ database/migrations/2018_12_14_120816_create_nilai_ekskul_table.php
@@ -18,7 +18,7 @@ class CreateNilaiEkskulTable extends Migration {
 			$table->string('id_nilai_ekskul', 40)->primary();
 			$table->string('id_pengambilan_ekskul', 40)->comment('FK: pengambilan_ekskul.id_pengambilan_ekskul');
 			$table->string('id_komponen_ekskul', 40)->comment('FK: komponen_ekskul.id_komponen_ekskul');
-			$table->float('besar_nilai_ekskul', 10, 0)->nullable();
+			$table->decimal('besar_nilai_ekskul', 10, 0)->nullable();
 			$table->timestamps();
 			$table->string('created_by', 40)->nullable();
 			$table->string('updated_by', 40)->nullable();
