+++ database/migrations/2019_04_18_231118_create_detail_biaya_internal_table.php
@@ -18,7 +18,7 @@ class CreateDetailBiayaInternalTable extends Migration
             $table->string('id_detail_biaya_internal', 40)->primary();
             $table->string('id_kelompok_biaya_internal', 40)->comment('FK: kelompok_biaya_internal.id_kelompok_biaya_internal');
             $table->string('nm_detail_biaya_internal', 128)->nullable();
-            $table->float('besar_biaya', 10, 0)->nullable();
+            $table->decimal('besar_biaya', 10, 0)->nullable();
             $table->timestamps();
             $table->string('created_by', 40)->nullable();
             $table->string('updated_by', 40)->nullable();
