+++ database/migrations/2019_09_16_123543_create_subkategori_pelanggaran_table.php
@@ -19,7 +19,7 @@ class CreateSubkategoriPelanggaranTable extends Migration
             $table->string('id_kategori_pelanggaran', 40)->comment('FK: kategori_pelanggaran.id_kategori_pelanggaran');
             $table->text('nm_subkategori_pelanggaran')->nullable();
             $table->integer('tingkat_subkategori_pelanggaran')->nullable();
-            $table->float('poin_subkategori_pelanggaran', 10, 0)->nullable();
+            $table->decimal('poin_subkategori_pelanggaran', 10, 0)->nullable();
             $table->string('keterangan_subkategori_pelanggaran', 512)->nullable();
             $table->timestamps();
             $table->string('created_by', 40)->nullable();
