+++ database/migrations/2019_12_09_114849_create_realisasi_pak_pembayaran_table.php
@@ -19,7 +19,7 @@ class CreateRealisasiPakPembayaranTable extends Migration
             $table->string('id_realisasi_pak', 40)->comment('FK: realisasi_pak.id_realisasi_pak');
             $table->boolean('termin_ke')->nullable();
             $table->date('tgl_pembayaran')->nullable();
-            $table->float('dana_realisasi_pembayaran', 10, 0)->nullable();
+            $table->decimal('dana_realisasi_pembayaran', 10, 0)->nullable();
             $table->string('id_pengguna_kepala_keuangan', 40)->nullable()->comment('FK: pengguna.id_pengguna, kepala keuangan yg melakukan approve Realisasi PAK Pembayaran');
             $table->timestamps();
             $table->string('created_by', 40)->nullable();
