+++ database/migrations/2020_02_19_113310_create_tutup_buku_bulanan_kas_table.php
@@ -18,10 +18,10 @@ class CreateTutupBukuBulananKasTable extends Migration
             $table->string('id_tutup_buku_bulanan_kas', 40)->primary();
             $table->string('id_semester', 40)->comment('FK: semester.id_semester');
             $table->boolean('id_bulan')->comment('1 = Januari, 2 = Februari, dst');
-            $table->float('kas_spp', 10, 0)->nullable();
-            $table->float('kas_rapb_penerimaan', 10, 0)->nullable();
-            $table->float('kas_rapb_pengeluaran', 10, 0)->nullable();
-            $table->float('kas_akhir_bulan', 10, 0)->nullable();
+            $table->decimal('kas_spp', 10, 0)->nullable();
+            $table->decimal('kas_rapb_penerimaan', 10, 0)->nullable();
+            $table->decimal('kas_rapb_pengeluaran', 10, 0)->nullable();
+            $table->decimal('kas_akhir_bulan', 10, 0)->nullable();
             $table->timestamps();
             $table->string('created_by', 40)->nullable();
             $table->string('updated_by', 40)->nullable();
