+++ database/migrations/2021_01_28_134020_add_urutan_to_rapor_kelompok_mp_and_rapor_subkelompok_mp.php
@@ -15,12 +15,12 @@ class AddUrutanToRaporKelompokMpAndRaporSubkelompokMp extends Migration
     {
         // add urutan to rapor_kelompok_mp
         Schema::table('rapor_kelompok_mp', function (Blueprint $table) {
-            $table->tinyInteger('urutan_rapor_kelompok_mp')->after('nm_rapor_kelompok_mp')->nullable();
+            $table->integer('urutan_rapor_kelompok_mp')->after('nm_rapor_kelompok_mp')->nullable();
         });
-        
+
         // add urutan to rapor_subkelompok_mp
         Schema::table('rapor_subkelompok_mp', function (Blueprint $table) {
-            $table->tinyInteger('urutan_rapor_subkelompok_mp')->after('id_mata_pelajaran')->nullable();
+            $table->integer('urutan_rapor_subkelompok_mp')->after('id_mata_pelajaran')->nullable();
         });
     }
