+++ database/migrations/2021_04_30_093333_add_column_jml_tunggakan_in_table_tb_bul_kas.php
@@ -14,7 +14,7 @@ class AddColumnJmlTunggakanInTableTbBulKas extends Migration
     public function up()
     {
         Schema::table('tutup_buku_bulanan_kas', function (Blueprint $table) {
-            $table->float('sisa_tunggakan_biaya', 10, 0)->nullable()->after('kas_akhir_bulan');
+            $table->decimal('sisa_tunggakan_biaya', 10, 0)->nullable()->after('kas_akhir_bulan');
         });
     }
