+++ database/migrations/2022_10_14_112830_add_stuf_in_table_komponen_nilai.php
@@ -14,7 +14,7 @@ class AddStufInTableKomponenNilai extends Migration
     public function up()
     {
         Schema::table('komponen_nilai_rapor_sisipan', function($table) {
-            $table->tinyInteger('status')->nullable()->after('urutan');
+            $table->integer('status')->nullable()->after('urutan');
             $table->string('type', 64)->nullable()->after('urutan');
         });
     }
