+++ database/migrations/2022_10_26_111648_change_default_to_number_besar_pembayaran.php
@@ -13,7 +13,7 @@ class ChangeDefaultToNumberBesarPembayaran extends Migration
     public function up()
     {
         Schema::table('tagihan_biaya', function ($table) {
-            $table->float('besar_pembayaran')->default('0')->change();
+            $table->decimal('besar_pembayaran', 10, 0)->default('0')->change();
         });
     }
